requirements.txt
numpy==1.24.3
scikit-learn==1.3.0
umap-learn==0.5.3
matplotlib==3.7.2 